requirements.txt
Flask
flask-bcrypt
flask-login
mysql-connector-python
jinja2
